feature: integrated new "MySegments" view refactor: renamed routes to be more relevant and created a dedicated "Home" view
