+++ src/routes/Segments/logo-wide.png
