+++ src/routes/index.tsx
@@ -1,3 +1,5 @@
 export { default as Edit } from './Edit';
 export { default as Home } from './Home';
-export { default as Search } from './Search';
+export { default as Videos } from './Videos';
+export { default as Segments } from './Segments';
+export { default as MySegments } from './MySegments';
