+++ src/utils/restful-resource-utility.tsx
@@ -5,12 +5,30 @@ export type Logger = typeof console;
 // TODO: Fill out params
 export type Params = {
   [key: string]: any;
-  $embed?: string | [string];
+  $embed?: string | string[];
+  $skip?: number;
+  $page?: number;
+  $limit?: number;
+  $sort?: string | string[];
 };
 
 export type ListResponse<T> = {
   data: {
-    docs: [T];
+    docs: T[];
+    pages: {
+      current: number;
+      prev: number;
+      hasPrev: boolean;
+      next: number;
+      hasNext: boolean;
+      total: number;
+    };
+    items: {
+      limit: number;
+      begin: number;
+      end: number;
+      total: number;
+    };
   };
 };
